+++ src/bin/readroa.rs
@@ -15,7 +15,7 @@ fn main() {
     let data = match fs::read(path) {
         Ok(data) => data,
         Err(err) => {
-            println!("Can’t read file: {}", err);
+            println!("Can’t read file: {err}");
             return;
         }
     };
@@ -23,7 +23,7 @@ fn main() {
     let _cert = match Roa::decode(data.as_ref(), true) {
         Ok(cert) => cert,
         Err(err) => {
-            println!("Can’t decode roa: {}", err);
+            println!("Can’t decode roa: {err}");
             return
         }
     };